Move report modelclasses into single file
