--- core/src/main/scala/stryker4s/report/model/Location.scala
@@ -1,3 +1,0 @@
-package stryker4s.report.model
-
-final case class Location(start: Position, end: Position)
