--- core/src/main/scala/stryker4s/report/model/MutantStatus.scala
@@ -1,6 +1,0 @@
-package stryker4s.report.model
-
-object MutantStatus extends Enumeration {
-  type MutantStatus = Value
-  val Killed, Survived, NoCoverage, CompileError, Timeout = Value
-}
