--- core/src/main/scala/stryker4s/report/model/MutationTestResult.scala
@@ -1,3 +1,0 @@
-package stryker4s.report.model
-
-final case class MutationTestResult(source: String, mutants: Seq[MutantResult], language: String = "scala")
