--- core/src/main/scala/stryker4s/report/model/Position.scala
@@ -1,3 +1,0 @@
-package stryker4s.report.model
-
-final case class Position(line: Int, column: Int)
