--- core/src/main/scala/stryker4s/report/model/Thresholds.scala
@@ -1,3 +1,0 @@
-package stryker4s.report.model
-
-final case class Thresholds(high: Int, low: Int)
